# shared/schemas/api.openapi.yaml
openapi: 3.0.3
info:
  title: Family Trivia API
  description: Backend API for the Family Trivia game
  version: 1.0.0
servers:
  - url: http://localhost:3001
    description: Local SAM development
paths:
...
  /submit-answer:
    post:
      summary: Submit an answer to a trivia question
      operationId: submitAnswer
      tags: [Trivia Questions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId, question, selectedAnswer]
              properties:
                userId:
                  type: string
                groupId:
                  type: string
                question:
                  $ref: "#/components/schemas/Question"
                selectedAnswer:
                  type: string
                isCatchingUp:
                  type: boolean
                isCustomCategory:
                  type: boolean
      responses:
        "200":
          description: Answer result
          content:
            application/json:
              schema:
                type: object
                properties:
                  correct:
                    type: boolean
                  streak:
                    type: integer
                  pointsEarned:
                    type: number
                  multiplier:
                    type: number
...
  /update-group-description:
    post:
      summary: Update a group description
      operationId: updateGroupDescription
      tags: [Family]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId, groupId, description]
              properties:
                userId:
                  type: string
                groupId:
                  type: string
                description:
                  type: string
      responses:
        "200":
          description: Update result
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  groupId:
                    type: string
                  description:
                    type: string
  /suggest-group-description:
    post:
      summary: Get AI-suggested group description
      operationId: suggestGroupDescription
      tags: [Family]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [groupId]
              properties:
                groupId:
                  type: string
                context:
                  type: string
      responses:
        "200":
          description: Suggested description
          content:
            application/json:
              schema:
                type: object
                properties:
                  suggestion:
                    type: string
...
components:
  schemas:
...
    DifficultyLevel:
      type: string
      enum: [easy, normal, hard]
...
    Question:
      type: object
      required: [question, choices, answer]
      properties:
        question:
          type: string
        choices:
          type: array
          items:
            type: string
        answer:
          type: string
        category:
          type: string
        difficulty:
          $ref: "#/components/schemas/DifficultyLevel"
        pointMultiplier:
          type: number
        mode:
          type: string
        type:
          type: string
        source:
          type: string
        sourceUrl:
          type: string
        externalId:
          type: string
...